2 player and deadzone camera
added physics, added 2 players. Player one is the girl and can be controlled with arrow keys and jump is the up arrow key. Player two is the robot boy and can be controlled with w,a, and d. There is also a deadzone camera that follows player one that will help a lot on the finished project.